+++ trql/src/lib.rs
@@ -2,8 +2,6 @@
 extern crate pest_derive;
 
 mod chain;
-pub mod command;
-pub mod parser;
 pub mod query;
 pub mod tree;
